// server/src/interfaces/error.ts
export enum ErrorTypes {
    TOO_MANY_REQUESTS = "TOO_MANY_REQUESTS",
    VALIDATION_ERR = "VALIDATION_ERR",
    INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
    MISSING_TOKEN = "MISSING_TOKEN",
    INVALID_TOKEN = "INVALID_TOKEN",
    ACCOUNT_LOCKED = "ACCOUNT_LOCKED",
    UNAUTHORIZED = "UNAUTHORIZED",
    NOT_FOUND = "NOT_FOUND",
    INTERNAL_ERR = "INTERNAL_ERR",
    VERSION_CONFLICT = "VERSION_CONFLICT",
    BAD_REQUEST = "BAD_REQUEST",
}